+++ src/main/java/net/minecraftforge/client/gui/config/ConfigElementList.java
@@ -1,7 +1,5 @@
 package net.minecraftforge.client.gui.config;
 
-import com.electronwill.nightconfig.core.UnmodifiableCommentedConfig;
-import com.electronwill.nightconfig.core.UnmodifiableConfig;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.IGuiEventListener;
@@ -15,18 +13,13 @@ import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.client.gui.config.ConfigElementControls.ConfigElementWidgetData;
-import net.minecraftforge.common.ForgeConfigSpec.*;
 import net.minecraftforge.fml.client.gui.widget.ExtendedButton;
 
-import javax.annotation.Nullable;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.function.BooleanSupplier;
-import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
@@ -36,169 +29,12 @@ public class ConfigElementList extends AbstractOptionList<ConfigElementList.Conf
     // From AbstractList#render/func_230430_a_
     private static final int SCROLLBAR_WIDTH = 6;
 
-    private final ConfigScreen configScreen;
+    protected final ConfigScreen configScreen;
     private int maxListLabelWidth;
 
     public ConfigElementList(ConfigScreen configScreen, Minecraft mcIn) {
         super(mcIn, configScreen.field_230708_k_, configScreen.field_230709_l_, 43, configScreen.field_230709_l_ - 32, 20);
         this.configScreen = configScreen;
-    }
-
-    public ConfigElementList(ConfigScreen configScreen, Minecraft mcIn, ConfigCategoryInfo info) {
-        this(configScreen, mcIn);
-        for (String key : info.elements())
-            this.func_230513_b_(createConfigElement(key, info));
-    }
-
-    protected ConfigElement createConfigElement(String key, ConfigCategoryInfo categoryInfo) {
-        Object raw = categoryInfo.getValue(key);
-        if (raw instanceof UnmodifiableConfig) {
-            UnmodifiableConfig value = (UnmodifiableConfig) raw;
-            UnmodifiableCommentedConfig valueInfo = (UnmodifiableCommentedConfig) categoryInfo.getSpec(key);
-            String description = categoryInfo.getCategoryComment(key);
-            ConfigCategoryInfo valueCategoryInfo = ConfigCategoryInfo.of(
-                    () -> value.valueMap().keySet(),
-                    value::get,
-                    valueInfo::get,
-                    valueInfo::getComment
-            );
-            return new CategoryConfigElement(configScreen, key, description == null ? "" : description, valueCategoryInfo);
-        } else {
-            ConfigValue<?> value = (ConfigValue<?>) raw;
-            ValueSpec valueInfo = (ValueSpec) categoryInfo.getSpec(key);
-            String name = new TranslationTextComponent(valueInfo.getTranslationKey()).getString();
-            Object valueValue = value.get();
-            ConfigElement element = createValueConfigElement(value, valueInfo, name, valueValue);
-            if (element != null)
-                return element;
-            return new TitledConfigElement(name + ": " + valueValue, valueInfo.getComment());
-        }
-    }
-
-    @Nullable
-    private ConfigElement createValueConfigElement(ConfigValue<?> value, ValueSpec valueInfo, String translatedName, Object valueValue) {
-        if (valueValue instanceof Boolean)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<Boolean> data = createBoolean((BooleanValue) value, translatedName, (Boolean) valueValue);
-                    canReset = data.canReset((BooleanValue) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof Integer)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<Integer> data = createNumericRanged((IntValue) value, valueInfo, translatedName, (Integer) valueValue, Integer::parseInt, Integer.MIN_VALUE);
-                    canReset = data.canReset((IntValue) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof Long)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<Long> data = createNumericRanged((LongValue) value, valueInfo, translatedName, (Long) valueValue, Long::parseLong, Long.MIN_VALUE);
-                    canReset = data.canReset((LongValue) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof Double)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<Double> data = createNumericRanged((DoubleValue) value, valueInfo, translatedName, (Double) valueValue, Double::parseDouble, Double.NEGATIVE_INFINITY);
-                    canReset = data.canReset((DoubleValue) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof Enum<?>)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<?> data = createEnum((EnumValue) value, valueInfo, translatedName, (Enum) valueValue);
-                    canReset = data.canReset((EnumValue) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof String)
-            return new TitledConfigElement(translatedName, valueInfo.getComment()) {
-                {
-                    ValueConfigElementData<String> data = createString((ConfigValue<String>) value, valueInfo, translatedName, (String) valueValue);
-                    canReset = data.canReset((ConfigValue<String>) value, valueInfo);
-                    reset = data.reset(valueInfo);
-                    widgets.add(0, data.widget);
-                }
-            };
-        if (valueValue instanceof List<?>)
-            return new PopupConfigElement(translatedName, valueInfo.getComment(), () -> new ListConfigScreen(configScreen, new StringTextComponent(translatedName), (List<?>) valueValue));
-        return null;
-    }
-
-    static class ValueConfigElementData<T> {
-        final Widget widget;
-        final Consumer<T> valueSetter;
-
-        ValueConfigElementData(Widget widget, Consumer<T> valueSetter) {
-            this.widget = widget;
-            this.valueSetter = valueSetter;
-        }
-
-        public Runnable reset(ValueSpec valueInfo) {
-            return () -> valueSetter.accept((T) valueInfo.getDefault());
-        }
-
-        public BooleanSupplier canReset(ConfigValue<T> value, ValueSpec valueInfo) {
-            return () -> !value.get().equals(valueInfo.getDefault());
-        }
-    }
-
-    public ValueConfigElementData<Boolean> createBoolean(BooleanValue value, String translatedName, Boolean valueValue) {
-        ConfigElementWidgetData<Boolean> control = ConfigElementControls.createBooleanButton(newValue -> {
-            value.set(newValue);
-            configScreen.onChange();
-            return true; // Can't have an "invalid" boolean
-        }, translatedName);
-        control.valueSetter.setup(valueValue);
-        return new ValueConfigElementData<>(control.widget, control.valueSetter::setTo);
-    }
-
-    public <T extends Comparable<? super T>> ValueConfigElementData<T> createNumericRanged(ConfigValue<T> value, ValueSpec valueInfo, String translatedName, T valueValue, Function<String, T> parser, T longestValue) {
-        @Nullable
-        Range<T> range = valueInfo.getRange();
-        ConfigElementWidgetData<T> control = ConfigElementControls.createNumericTextField(newValue -> {
-            if (range != null && !range.test(newValue))
-                return false;
-            value.set(newValue);
-            configScreen.onChange();
-            return true;
-        }, translatedName, parser, longestValue);
-        control.valueSetter.setup(valueValue);
-        return new ValueConfigElementData<>(control.widget, control.valueSetter::setTo);
-    }
-
-    private <T extends Enum<T>> ValueConfigElementData<T> createEnum(EnumValue<T> value, ValueSpec valueInfo, String translatedName, T valueValue) {
-        Enum<?>[] potential = Arrays.stream(((Enum<?>) valueValue).getDeclaringClass().getEnumConstants()).filter(valueInfo::test).toArray(Enum<?>[]::new);
-        ConfigElementWidgetData<T> control = ConfigElementControls.createEnumButton(newValue -> {
-            value.set(newValue);
-            configScreen.onChange();
-            return true; // Can't have an "invalid" boolean
-        }, translatedName, (T[]) potential);
-        control.valueSetter.setup(valueValue);
-        return new ValueConfigElementData<>(control.widget, control.valueSetter::setTo);
-    }
-
-    private ValueConfigElementData<String> createString(ConfigValue<String> value, ValueSpec valueInfo, String translatedName, String valueValue) {
-        ConfigElementWidgetData<String> control = ConfigElementControls.createStringTextField(newValue -> {
-            if (!valueInfo.test(newValue))
-                return false;
-            value.set(newValue);
-            configScreen.onChange();
-            return true;
-        }, translatedName);
-        control.valueSetter.setup(valueValue);
-        return new ValueConfigElementData<>(control.widget, control.valueSetter::setTo);
     }
 
     @Override
@@ -372,9 +208,4 @@ public class ConfigElementList extends AbstractOptionList<ConfigElementList.Conf
         }
     }
 
-    public class CategoryConfigElement extends PopupConfigElement {
-        public CategoryConfigElement(ConfigScreen parentScreen, String translatedName, String description, ConfigCategoryInfo valueCategoryInfo) {
-            super(translatedName, description, () -> new CategoryConfigScreen(parentScreen, new StringTextComponent(translatedName), valueCategoryInfo));
-        }
-    }
 }
